app/requirements.txt
pyinstaller
pycryptodome
